Add plugin for Piwik web analytics
This is a plugin that adds the javascript tracking code for the
use of the open source web analytics tool Piwik <http://piwik.org/>
to the page header.

For the configuration, two parameters are needed:
baseUrl: The URL of an existing Piwik installation, without http(s)://
siteId: The ID of the site that is set up for tracking in Piwik.

This is heavily based on the existing Google Analytics plugin.