PiwikPlugin should also track modal photo views
Until now, the Piwik tracker code inserted by the PiwikPlugin would
only do tracking at page load, thus missing any modal AJAX photo views.

This adds a piece of Javascript to the PiwikPlugin, which installs
a listener on the custom event that gets fired when a modal photo
link is clicked.

The updated page title and URL are then reported to the Piwik tracker
as an additional page view.