Fixing a bug where the create tag API would always return a 500 on success. Closes #927
There was a lot wrong with the create tag API. It's not actually used in any code I know of so it seems it went undetected for a while (need unit tests).

Reproducing the 500 issue.

    ./openphoto -X POST -h 2.local.photo -e '/tag/create.json' -F 'tag=foobar' -p
    {
      "message" : "An unknown error occurred.",
      "code" : 500,
      "result" : null
    }

From the apache logs.

    {severity:warn, description:"Uncaught exception (/vagrant/frontend/src/libraries/external/epi/EpiException.php:13): Could not find route (/v1/tag/foobar/view.json) from (/tag/create.json?oauth_consumer_key=ea63f99a0f3f4498121c0b3cbac5b4)", additional:}

Calling the newly created `foobar` tag which has a count of 0 shoulr return a 404 `/tag/foobar/view.json` but we get HTML.

    ./openphoto -h 2.local.photo -e '/tag/foobar/view.json' -p
    <!DOCTYPE html>
    <html lang="en">
      <head>
    [snip]

After the fix.

The create API should return a 201.

    ./openphoto -X POST -h 2.local.photo -e '/tag/create.json' -F 'tag=foobar2' -p
    {
      "message" : "Tag foobar2 created successfully.",
      "code" : 201,
      "result" : true
    }

Querying it should return a 404 if no photos are tagged with this tag or a 200 if they are.

    ./openphoto -h 2.local.photo -e '/tag/foobar/view.json' -p
    {
      "message" : "Could not find tag foobar",
      "code" : 404,
      "result" : false
    }

    ./openphoto -h 2.local.photo -e '/tag/foobar2/view.json' -p
    {
      "message" : "Successfully returned tag foobar2",
      "code" : 200,
      "result" : {
        "id" : "foobar2",
        "owner" : "2@jmathai.com",
        "count" : 1
      }
    }